+++ install-build-dependencies.sh
@@ -10,8 +10,4 @@ sudo apt-get install -y --no-install-recommends \
   inkscape
 
 python -m pip install -U pip
-python -m pip install --upgrade-strategy=eager \
-  wheel \
-  pynsist==2.8 \
-  distlib==0.3.3 \
-  freezegun
+python -m pip install -U --upgrade-strategy=eager -r requirements.txt